chore: ignore pyup updates for packages that will break python 3.7
To allow devs to develop on pure python 3.7 environment, ignore
packages (used for development purposes) are not supported
in python 3.7.

An alternative is to suggest devs to work only in python 3.11,
even though the library itself can be used in python 3.7-3.11.

We will consider the alternatives if there are compelling
reasons to use newer versions of those packages during
development process.